Ejercicio 3 terminado

+++ app/Http/Controllers/PostController.php
@@ -37,6 +37,7 @@ class PostController extends Controller
     {
         $data = $request->validated();
         $data['slug'] = Str::slug($data['title']);
+        $data['summary'] = Str::limit($data['body'], 50);
 
         auth()->user()->posts()->create($data);
 
@@ -51,7 +52,9 @@ class PostController extends Controller
 
     public function update(UpdatePostRequest $request, Post $post)
     {
-        $post->update($request->validated());
+        $data = $request->validated();
+
+        $post->update($data);
 
         return to_route('posts.show', $post)
             ->with('status', 'Post updated successfully');
@@ -59,6 +62,10 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        if (!in_array($post->status, ['draft', 'pending'])) {
+            return redirect()->route('posts.index')->with('error', 'Solo se pueden eliminar posts con estado "borrador" o "pendiente".');
+        }
+
         $post->delete();
 
         return to_route('posts.index')
